requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Flask==2.1.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.0
MarkupSafe==2.1.1
numpy==1.22.4
pandas==1.4.2
pathlib==1.0.1
Pillow==9.1.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
requests==2.28.0
six==1.16.0
typing==3.7.4.3
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
